+++ apps/Standalone/src/designer/app/AzureLogicAppsDesigner/DesignerCommandBar.tsx
@@ -130,7 +130,9 @@ export const DesignerCommandBar = ({
 
     if (!hasParametersErrors) {
       await saveWorkflow(serializedWorkflow, customCodeFilesWithData, () => dispatch(resetDesignerDirtyState(undefined)));
-      updateCallbackUrl(designerState, DesignerStore.dispatch);
+      if (Object.keys(serializedWorkflow?.definition?.triggers ?? {}).length > 0) {
+        updateCallbackUrl(designerState, dispatch);
+      }
     }
   });
 
@@ -415,18 +417,6 @@ export const DesignerCommandBar = ({
 
   return (
     <>
-      <div
-        style={{
-          position: 'relative',
-          top: '60px',
-          left: '20px',
-          display: 'flex',
-          flexDirection: 'column',
-          gap: '8px',
-        }}
-      >
-        {!isInitialized && <Spinner size={'extra-small'} label={'Loading dynamic data...'} />}
-      </div>
       <CommandBar
         items={isMonitoringView ? monitoringItems : editorItems}
         ariaLabel="Use left and right arrow keys to navigate between commands"
@@ -438,6 +428,19 @@ export const DesignerCommandBar = ({
           },
         }}
       />
+      <div
+        style={{
+          position: 'relative',
+          top: '60px',
+          left: '20px',
+          display: 'flex',
+          flexDirection: 'column',
+          gap: '8px',
+          height: '0px',
+        }}
+      >
+        {!isInitialized && <Spinner size={'extra-small'} label={'Loading dynamic data...'} />}
+      </div>
     </>
   );
 };