+++ apps/Standalone/src/designer/app/SettingsSections/contextSettings.tsx
@@ -9,6 +9,7 @@ import {
   useShowPerformanceDebug,
   useSuppressDefaultNodeSelect,
   useStringOverrides,
+  useQueryCachePersist,
 } from '../../state/workflowLoadingSelectors';
 import {
   setDarkMode,
@@ -21,6 +22,7 @@ import {
   setShowPerformanceDebug,
   setSuppressDefaultNodeSelect,
   setStringOverrides,
+  setQueryCachePersist,
 } from '../../state/workflowLoadingSlice';
 import { Checkbox, TextField } from '@fluentui/react';
 import { useCallback } from 'react';
@@ -36,6 +38,7 @@ const ContextSettings = () => {
   const hostOptions = useHostOptions();
   const showPerformanceDebug = useShowPerformanceDebug();
   const showTestStringOverride = useStringOverrides();
+  const queryCachePersist = useQueryCachePersist();
   const dispatch = useDispatch<AppDispatch>();
 
   const changeMonitoringView = useCallback(
@@ -103,6 +106,11 @@ const ContextSettings = () => {
             )
           );
         }}
+      />
+      <Checkbox
+        label="Query Cache Persist"
+        checked={queryCachePersist}
+        onChange={(_, checked) => dispatch(setQueryCachePersist(!!checked))}
       />
       <TextField
         label="Max State History Size"