+++ libs/designer-ui/src/lib/authentication/index.tsx
@@ -82,11 +82,7 @@ export const AuthenticationEditor = ({
   options,
   authenticationValue,
   initialValue,
-  getTokenPicker,
   onChange,
-  readonly,
-  tokenMapping,
-  loadParameterValueFromString,
   ...props
 }: AuthenticationEditorProps): JSX.Element => {
   const intl = useIntl();
@@ -120,64 +116,34 @@ export const AuthenticationEditor = ({
       case AuthenticationType.BASIC:
         return (
           <BasicAuthentication
+            {...props}
             basicProps={basic}
             tokenPickerButtonProps={props.tokenPickerButtonProps}
-            readonly={readonly}
             setCurrentProps={setCurrentProps}
-            getTokenPicker={getTokenPicker}
-            tokenMapping={tokenMapping}
-            loadParameterValueFromString={loadParameterValueFromString}
           />
         );
       case AuthenticationType.CERTIFICATE:
         return (
           <CertificateAuthentication
+            {...props}
             clientCertificateProps={clientCertificate}
             tokenPickerButtonProps={props.tokenPickerButtonProps}
-            readonly={readonly}
             setCurrentProps={setCurrentProps}
-            getTokenPicker={getTokenPicker}
-            tokenMapping={tokenMapping}
-            loadParameterValueFromString={loadParameterValueFromString}
           />
         );
       case AuthenticationType.RAW:
         return (
           <RawAuthentication
+            {...props}
             rawProps={raw}
             tokenPickerButtonProps={props.tokenPickerButtonProps}
-            readonly={readonly}
-            getTokenPicker={getTokenPicker}
             setCurrentProps={setCurrentProps}
-            tokenMapping={tokenMapping}
-            loadParameterValueFromString={loadParameterValueFromString}
           />
         );
       case AuthenticationType.MSI:
-        return (
-          <MSIAuthentication
-            identity={options?.identity}
-            msiProps={msi}
-            readonly={readonly}
-            tokenPickerButtonProps={props.tokenPickerButtonProps}
-            setCurrentProps={setCurrentProps}
-            getTokenPicker={getTokenPicker}
-            tokenMapping={tokenMapping}
-            loadParameterValueFromString={loadParameterValueFromString}
-          />
-        );
+        return <MSIAuthentication {...props} identity={options?.identity} msiProps={msi} setCurrentProps={setCurrentProps} />;
       case AuthenticationType.OAUTH:
-        return (
-          <ActiveDirectoryAuthentication
-            OauthProps={aadOAuth}
-            readonly={readonly}
-            tokenPickerButtonProps={props.tokenPickerButtonProps}
-            setCurrentProps={setCurrentProps}
-            getTokenPicker={getTokenPicker}
-            tokenMapping={tokenMapping}
-            loadParameterValueFromString={loadParameterValueFromString}
-          />
-        );
+        return <ActiveDirectoryAuthentication OauthProps={aadOAuth} setCurrentProps={setCurrentProps} />;
       case AuthenticationType.NONE:
         return null;
       default:
@@ -218,7 +184,7 @@ export const AuthenticationEditor = ({
       {expandedView ? (
         <div className="msla-authentication-editor-expanded-container">
           <AuthenticationDropdown
-            readonly={readonly}
+            readonly={props.readonly}
             dropdownLabel={authenticationTypeLabel}
             selectedKey={option}
             options={getAuthenticationTypes(options.supportedAuthTypes)}
@@ -229,15 +195,12 @@ export const AuthenticationEditor = ({
       ) : (
         <>
           <CollapsedAuthentication
+            {...props}
             collapsedValue={collapsedValue}
             setErrorMessage={setCollapsedErrorMessage}
             setCurrentProps={setCurrentProps}
             setOption={setOption}
             serializeValue={serializeCodeCollapsedValue}
-            readonly={readonly}
-            getTokenPicker={getTokenPicker}
-            tokenMapping={tokenMapping}
-            loadParameterValueFromString={loadParameterValueFromString}
           />
           <div className="msla-auth-editor-validation">{collapsedErrorMessage}</div>
         </>