+++ libs/designer/src/lib/core/templates/TemplatesDataProvider.tsx
@@ -6,7 +6,6 @@ import type { AppDispatch, RootState } from '../state/templates/store';
 import {
   loadGithubManifestNames,
   loadGithubManifests,
-  setCustomTemplates,
   setFilteredTemplateNames,
   templatesCountPerPage,
   lazyLoadGithubManifests,
@@ -15,8 +14,8 @@ import { type ResourceDetails, setInitialData } from '../state/templates/workflo
 import { useAreServicesInitialized } from '../state/templates/templateselectors';
 import type { ConnectionReferences } from '../../common/models/workflow';
 import { getFilteredTemplates } from './utils/helper';
-import { initializeTemplateServices } from '../actions/bjsworkflow/templates';
-import type { Template } from '@microsoft/logic-apps-shared';
+import { initializeTemplateServices, reloadTemplates } from '../actions/bjsworkflow/templates';
+import { InitTemplateService, type Template } from '@microsoft/logic-apps-shared';
 import { setViewTemplateDetails } from '../state/templates/templateOptionsSlice';
 import { changeCurrentTemplateName } from '../state/templates/templateSlice';
 
@@ -27,18 +26,33 @@ export interface TemplatesDataProviderProps {
   resourceDetails: ResourceDetails;
   services: TemplateServiceOptions;
   connectionReferences: ConnectionReferences;
-  customTemplates?: Record<string, Template.Manifest>;
   viewTemplate?: Template.ViewTemplateDetails;
   children?: React.ReactNode;
+  reload?: boolean;
 }
 
-const DataProviderInner = ({ isConsumption, children }: TemplatesDataProviderProps) => {
+const DataProviderInner = ({ isConsumption, children, reload, services }: TemplatesDataProviderProps) => {
   const dispatch = useDispatch<AppDispatch>();
-  const { githubTemplateNames, availableTemplates, filters } = useSelector((state: RootState) => state?.manifest);
+  const { githubTemplateNames, availableTemplates, filters, servicesInitialized } = useSelector((state: RootState) => ({
+    githubTemplateNames: state.manifest.githubTemplateNames,
+    availableTemplates: state.manifest.availableTemplates,
+    filters: state.manifest.filters,
+    servicesInitialized: state.templateOptions.servicesInitialized,
+  }));
 
   useEffect(() => {
     dispatch(loadGithubManifestNames());
   }, [dispatch]);
+
+  useEffect(() => {
+    if (reload !== undefined) {
+      if (servicesInitialized && services.templateService) {
+        InitTemplateService(services.templateService);
+      }
+      console.log('reloadTemplates got called');
+      dispatch(reloadTemplates({ clear: true }));
+    }
+  }, [reload, dispatch, services.templateService, servicesInitialized]);
 
   useEffect(() => {
     if (githubTemplateNames) {
@@ -105,12 +119,6 @@ export const TemplatesDataProvider = (props: TemplatesDataProviderProps) => {
       dispatch(setViewTemplateDetails(props.viewTemplate));
     }
   }, [dispatch, props.viewTemplate]);
-
-  useEffect(() => {
-    if (props.customTemplates) {
-      dispatch(setCustomTemplates(props.customTemplates));
-    }
-  }, [dispatch, props.customTemplates]);
 
   if (!servicesInitialized) {
     return null;