+++ libs/designer/src/lib/core/templates/TemplatesDesignerContext.tsx
@@ -1,9 +1,6 @@
 import type {
-  IApiManagementService,
-  IAppServiceService,
   IConnectionParameterEditorService,
   IConnectionService,
-  IFunctionService,
   IGatewayService,
   ITenantService,
   ILoggerService,
@@ -13,6 +10,7 @@ import type {
   IOperationManifestService,
   IDesignerUiInteractionsService,
   IExperimentationService,
+  IConnectorService,
 } from '@microsoft/logic-apps-shared';
 import { createContext } from 'react';
 
@@ -23,14 +21,12 @@ export interface TemplatesDesignerContext {
 export interface TemplateServiceOptions {
   connectionService: IConnectionService;
   operationManifestService: IOperationManifestService;
+  connectorService?: IConnectorService;
   workflowService: IWorkflowService;
   gatewayService?: IGatewayService;
   tenantService?: ITenantService;
   loggerService?: ILoggerService;
   oAuthService: IOAuthService;
-  apimService?: IApiManagementService;
-  functionService?: IFunctionService;
-  appServiceService?: IAppServiceService;
   connectionParameterEditorService?: IConnectionParameterEditorService;
   templateService?: ITemplateService;
   uiInteractionsService?: IDesignerUiInteractionsService;